requirements.txt
attrs==19.3.0
Django==2.2
django-debug-toolbar==2.2
django-extensions==2.2.9
dummycache==0.0.2
graphviz==0.14
isort==5.2.0
mock==4.0.2
more-itertools==8.4.0
packaging==20.4
Pillow==7.2.0
pluggy==0.13.1
py==1.8.2
pydotplus==2.0.2
pyparsing==2.4.7
pytest==5.4.3
pytest-django==3.8.0
pytz==2020.1
six==1.15.0
sorl-thumbnail==12.6.3
sqlparse==0.3.1
wcwidth==0.2.5
